fix: add security checks for deletion
